LayoutTests/platform/blackberry/fast/dom/location-hash-expected.txt

This tests that modifying location.hash works as it should

PASS: window.history.length == originalHistoryLength should be true and is.
PASS: window.location.hash should be and is.
PASS: window.location.hash should be #foo and is.
PASS: window.location == originalLocation + '#foo' should be true and is.
FAIL: window.history.length == originalHistoryLength + 1 should be true but instead is false.
PASS: window.location.hash should be #bar and is.
PASS: window.location == originalLocation + '#bar' should be true and is.
FAIL: window.history.length == originalHistoryLength + 2 should be true but instead is false.
PASS: window.location.hash should be #bar and is.
PASS: window.location == originalLocation + '#bar' should be true and is.
FAIL: window.history.length == originalHistoryLength + 2 should be true but instead is false.
PASS: window.location.hash should be #bar and is.
PASS: window.location == originalLocation + '#bar' should be true and is.
PASS: window.location.hash should be #bar and is.
PASS: window.location == originalLocation + '#bar' should be true and is.
PASS: window.location.hash should be #foo and is.
PASS: window.location == originalLocation + '#foo' should be true and is.
PASS: window.location.hash should be #foo and is.
PASS: window.location == originalLocation + '#foo' should be true and is.
PASS: window.location.hash should be and is.
PASS: window.location == originalLocation should be true and is.
PASS: window.location.hash should be and is.
PASS: window.location == originalLocation should be true and is.
PASS: window.location.hash should be #foo and is.
PASS: window.location == originalLocation + '#foo' should be true and is.
FAILURE: one or more tests failed
